requirements.txt
aiobotocore==2.19.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aioitertools==0.12.0
aiosignal==1.3.1
airbyte==0.16.4
airbyte-api==0.49.4
airbyte-cdk==3.0.0
airbyte_protocol_models_pdv2==0.12.2
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
botocore==1.36.3
bracex==2.5.post1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
dpath==2.2.0
duckdb==1.1.3
duckdb_engine==0.13.6
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.21
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.12.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
genson==1.2.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==4.44.1
gradio_client==1.3.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.3
grpcio-status==1.64.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==0.2
jsonschema==3.2.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.1.42
langchain-experimental==0.3.3
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
layoutparser==0.3.4
llama_stack_client==0.1.5
lxml==5.3.0
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.4
mdurl==0.1.2
mmhash3==3.0.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.55.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pendulum==2.1.2
pi_heif==0.21.0
pikepdf==9.5.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.1
psycopg2-binary==2.9.9
pyaml==25.1.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyiceberg==0.6.1
PyJWT==2.10.1
pymongo==4.13.0
pyOpenSSL==24.3.0
pypandoc==1.15
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.1
pyrate-limiter==3.1.1
pyrsistent==0.20.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
python-ulid==2.7.0
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruff==0.11.7
s3fs==2024.12.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.4
snowflake-sqlalchemy==1.7.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.51
sqlalchemy-bigquery==1.9.0
starlette==0.45.2
streamlit==1.40.2
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
together==1.4.6
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.2
types-PyYAML==6.0.12.20241230
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.15.14
unstructured-client==0.29.0
unstructured-inference==0.7.36
unstructured-ingest==0.3.0
unstructured.pytesseract==0.3.13
url-normalize==1.4.3
urllib3==2.4.0
uuid7==0.1.0
uvicorn==0.34.0
wcmatch==8.4
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.1
yarl==1.18.0
zipp==3.21.0
